Split out YamlRenderer into its own file, and clean it up a bit
